refactor(bot): split errorhandlers module into focused units
Split monolithic errorHandlers.ts (418 lines) into four focused
modules with strict one-way import discipline:
- errorClassification.ts: pure error helpers, no cross-module imports
- streamFailureNotifier.ts: Discord channel notifications
- streamRecovery.ts: YouTube parser and stream extraction recovery
- errorEventHandlers.ts: player event setup and wiring

Correspondingly split spec into 4 focused unit test suites. All 233
tests pass. Mutation test confirmed effectiveness (isSameTrack
boundary condition caught by tests).
