--- packages/bot/src/handlers/player/errorHandlers.spec.ts
@@ -1,803 +1,0 @@
-import { beforeEach, describe, expect, it, jest } from '@jest/globals'
-import { setupErrorHandlers } from './errorHandlers'
-
-const debugLogMock = jest.fn()
-const errorLogMock = jest.fn()
-const warnLogMock = jest.fn()
-const captureExceptionMock = jest.fn()
-const analyzeYouTubeErrorMock = jest.fn()
-const logYouTubeErrorMock = jest.fn()
-const recordFailureMock = jest.fn()
-const recordSuccessMock = jest.fn()
-const providerFromTrackMock = jest.fn()
-const createErrorEmbedMock = jest.fn(() => ({ type: 'error' }))
-
-jest.mock('@lucky/shared/utils', () => ({
-    debugLog: (...args: unknown[]) => debugLogMock(...args),
-    errorLog: (...args: unknown[]) => errorLogMock(...args),
-    warnLog: (...args: unknown[]) => warnLogMock(...args),
-    captureException: (...args: unknown[]) => captureExceptionMock(...args),
-    sanitizeErrorMessage: (e: unknown) =>
-        e instanceof Error ? e.message : String(e),
-    sanitizeStack: (e: unknown) => (e instanceof Error ? e.stack : undefined),
-}))
-
-jest.mock('../../utils/general/embeds', () => ({
-    createErrorEmbed: (...args: unknown[]) => createErrorEmbedMock(...args),
-}))
-
-jest.mock('../../utils/music/youtubeErrorHandler', () => ({
-    analyzeYouTubeError: (...args: unknown[]) =>
-        analyzeYouTubeErrorMock(...args),
-    logYouTubeError: (...args: unknown[]) => logYouTubeErrorMock(...args),
-}))
-
-jest.mock('@lucky/shared/config', () => ({
-    youtubeConfig: {
-        errorHandling: { skipOnParserError: true },
-    },
-}))
-
-jest.mock('../../utils/music/search/providerHealth', () => ({
-    providerFromTrack: (...args: unknown[]) => providerFromTrackMock(...args),
-    providerHealthService: {
-        recordFailure: (...args: unknown[]) => recordFailureMock(...args),
-        recordSuccess: (...args: unknown[]) => recordSuccessMock(...args),
-    },
-}))
-
-type QueueErrorHandler = (queue: any, error: Error) => void
-type PlayerErrorHandler = (queue: any, error: Error, track?: any) => unknown
-type DebugHandler = (queue: any, message: string) => void
-type TopLevelErrorHandler = (error: Error) => void
-type TopLevelDebugHandler = (message: string) => void
-
-async function flushPromises(): Promise<void> {
-    await new Promise<void>((resolve) => {
-        setImmediate(() => resolve())
-    })
-}
-
-function createPlayerWithHandlers(): {
-    queueHandlers: Record<
-        string,
-        QueueErrorHandler | PlayerErrorHandler | DebugHandler
-    >
-    playerHandlers: Record<string, TopLevelErrorHandler | TopLevelDebugHandler>
-} {
-    const queueHandlers: Record<
-        string,
-        QueueErrorHandler | PlayerErrorHandler | DebugHandler
-    > = {}
-    const playerHandlers: Record<
-        string,
-        TopLevelErrorHandler | TopLevelDebugHandler
-    > = {}
-    const player = {
-        events: {
-            on: jest.fn(
-                (
-                    event: string,
-                    handler:
-                        QueueErrorHandler | PlayerErrorHandler | DebugHandler,
-                ) => {
-                    queueHandlers[event] = handler
-                },
-            ),
-        },
-        on: jest.fn(
-            (
-                event: string,
-                handler: TopLevelErrorHandler | TopLevelDebugHandler,
-            ) => {
-                playerHandlers[event] = handler
-            },
-        ),
-    }
-
-    setupErrorHandlers(player as any)
-    return { queueHandlers, playerHandlers }
-}
-
-describe('setupErrorHandlers', () => {
-    beforeEach(() => {
-        jest.clearAllMocks()
-        analyzeYouTubeErrorMock.mockReturnValue({
-            isParserError: false,
-            isCompositeVideoError: false,
-            isHypePointsError: false,
-            isTypeMismatchError: false,
-        })
-        providerFromTrackMock.mockReturnValue('youtube')
-    })
-
-    it('records provider failure and recovers stream extraction with alternative track', async () => {
-        const queueHandlers: Record<
-            string,
-            QueueErrorHandler | PlayerErrorHandler | DebugHandler
-        > = {}
-        const playerHandlers: Record<
-            string,
-            TopLevelErrorHandler | TopLevelDebugHandler
-        > = {}
-        const player = {
-            events: {
-                on: jest.fn(
-                    (
-                        event: string,
-                        handler:
-                            | QueueErrorHandler
-                            | PlayerErrorHandler
-                            | DebugHandler,
-                    ) => {
-                        queueHandlers[event] = handler
-                    },
-                ),
-            },
-            on: jest.fn(
-                (
-                    event: string,
-                    handler: TopLevelErrorHandler | TopLevelDebugHandler,
-                ) => {
-                    playerHandlers[event] = handler
-                },
-            ),
-        }
-        setupErrorHandlers(player as any)
-
-        const alternativeTrack = { url: 'https://example.com/alt' }
-        const queue = {
-            guild: { name: 'Guild 1' },
-            metadata: { requestedBy: { id: 'user-1' } },
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Song A',
-                requestedBy: { id: 'user-1' },
-            },
-            player: {
-                search: jest.fn().mockResolvedValue({
-                    tracks: [alternativeTrack],
-                }),
-            },
-            insertTrack: jest.fn(),
-            node: {
-                skip: jest.fn(),
-            },
-        }
-
-        expect(playerHandlers.error).toEqual(expect.any(Function))
-        expect(playerHandlers.debug).toEqual(expect.any(Function))
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(recordFailureMock).toHaveBeenCalledWith(
-            'youtube',
-            expect.any(Number),
-            'Could not extract stream',
-        )
-        expect(queue.player.search).toHaveBeenCalled()
-        expect(queue.insertTrack).toHaveBeenCalledWith(alternativeTrack, 0)
-        expect(queue.node.skip).toHaveBeenCalled()
-        expect(recordSuccessMock).toHaveBeenCalledWith('youtube')
-    })
-
-    it('skips without reinserting when alternative track has same URL', async () => {
-        const queueHandlers: Record<
-            string,
-            QueueErrorHandler | PlayerErrorHandler | DebugHandler
-        > = {}
-        const playerHandlers: Record<
-            string,
-            TopLevelErrorHandler | TopLevelDebugHandler
-        > = {}
-        const player = {
-            events: {
-                on: jest.fn(
-                    (
-                        event: string,
-                        handler:
-                            | QueueErrorHandler
-                            | PlayerErrorHandler
-                            | DebugHandler,
-                    ) => {
-                        queueHandlers[event] = handler
-                    },
-                ),
-            },
-            on: jest.fn(
-                (
-                    event: string,
-                    handler: TopLevelErrorHandler | TopLevelDebugHandler,
-                ) => {
-                    playerHandlers[event] = handler
-                },
-            ),
-        }
-        setupErrorHandlers(player as any)
-
-        const queue = {
-            guild: { name: 'Guild 1', id: 'guild-1' },
-            metadata: { requestedBy: { id: 'user-1' }, channel: null },
-            currentTrack: {
-                url: 'https://www.youtube.com/watch?v=abc123',
-                title: 'Song A',
-                requestedBy: { id: 'user-1' },
-            },
-            player: {
-                search: jest.fn().mockResolvedValue({
-                    tracks: [
-                        {
-                            url: 'https://www.youtube.com/watch?v=abc123',
-                            title: 'Song A',
-                        },
-                    ],
-                }),
-            },
-            insertTrack: jest.fn(),
-            node: {
-                skip: jest.fn(),
-            },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(queue.player.search).toHaveBeenCalled()
-        expect(queue.insertTrack).not.toHaveBeenCalled()
-        expect(queue.node.skip).toHaveBeenCalled()
-        expect(recordSuccessMock).not.toHaveBeenCalled()
-        expect(warnLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: expect.stringContaining('same track alternative'),
-            }),
-        )
-    })
-
-    it('skips track on parser errors when skip config is enabled', async () => {
-        const queueHandlers: Record<
-            string,
-            QueueErrorHandler | PlayerErrorHandler | DebugHandler
-        > = {}
-        const playerHandlers: Record<
-            string,
-            TopLevelErrorHandler | TopLevelDebugHandler
-        > = {}
-        const player = {
-            events: {
-                on: jest.fn(
-                    (
-                        event: string,
-                        handler:
-                            | QueueErrorHandler
-                            | PlayerErrorHandler
-                            | DebugHandler,
-                    ) => {
-                        queueHandlers[event] = handler
-                    },
-                ),
-            },
-            on: jest.fn(
-                (
-                    event: string,
-                    handler: TopLevelErrorHandler | TopLevelDebugHandler,
-                ) => {
-                    playerHandlers[event] = handler
-                },
-            ),
-        }
-        setupErrorHandlers(player as any)
-        analyzeYouTubeErrorMock.mockReturnValue({
-            isParserError: true,
-            isCompositeVideoError: false,
-            isHypePointsError: false,
-            isTypeMismatchError: true,
-        })
-
-        const queue = {
-            guild: { name: 'Guild 2' },
-            metadata: { requestedBy: { id: 'user-2' } },
-            currentTrack: {
-                url: 'https://example.com/current',
-                requestedBy: { id: 'user-2' },
-            },
-            node: { skip: jest.fn() },
-        }
-
-        expect(playerHandlers.error).toEqual(expect.any(Function))
-        expect(playerHandlers.debug).toEqual(expect.any(Function))
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('parser failed'),
-        )
-        await flushPromises()
-
-        expect(logYouTubeErrorMock).toHaveBeenCalled()
-        expect(queue.node.skip).toHaveBeenCalled()
-    })
-
-    it('handles top-level player errors and queue errors without throwing', () => {
-        const { queueHandlers, playerHandlers } = createPlayerWithHandlers()
-
-        const queue = {
-            guild: { id: 'guild-1', name: 'Guild 1' },
-            connection: {
-                state: { status: 'disconnected' },
-                rejoin: jest.fn(() => {
-                    throw new Error('rejoin failed')
-                }),
-            },
-        }
-
-        expect(() =>
-            (queueHandlers.error as QueueErrorHandler)(
-                queue,
-                new Error('ECONNRESET test'),
-            ),
-        ).not.toThrow()
-        expect(() =>
-            (playerHandlers.error as TopLevelErrorHandler)(
-                new Error('Unhandled player error'),
-            ),
-        ).not.toThrow()
-
-        expect(errorLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: 'Unhandled player error:',
-                data: expect.objectContaining({
-                    errorMessage: 'Unhandled player error',
-                }),
-            }),
-        )
-        expect(errorLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: 'Error in queue Guild 1:',
-                data: expect.objectContaining({
-                    guildId: 'guild-1',
-                    errorMessage: 'ECONNRESET test',
-                }),
-            }),
-        )
-        // Both top-level and queue player errors must reach Sentry.
-        expect(captureExceptionMock).toHaveBeenCalledWith(
-            expect.any(Error),
-            expect.objectContaining({ context: 'player-unhandled-error' }),
-        )
-        expect(captureExceptionMock).toHaveBeenCalledWith(
-            expect.any(Error),
-            expect.objectContaining({
-                context: 'player-queue-error',
-                guildId: 'guild-1',
-            }),
-        )
-    })
-
-    it('normalizes non-error queue payloads and recovers connection when rejoin works', () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        const queue = {
-            guild: { id: 'guild-raw', name: 'Guild Raw' },
-            connection: {
-                state: { status: 'disconnected' },
-                rejoin: jest.fn(),
-            },
-        }
-
-        ;(queueHandlers.error as QueueErrorHandler)(
-            queue as any,
-            'ECONNRESET' as any,
-        )
-
-        expect(queue.connection.rejoin).toHaveBeenCalled()
-        expect(errorLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: 'Error in queue Guild Raw:',
-                data: expect.objectContaining({
-                    errorName: 'string',
-                    errorMessage: 'ECONNRESET',
-                }),
-            }),
-        )
-        expect(debugLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: 'Attempting to recover from connection error',
-            }),
-        )
-        // Non-Error queue payload is wrapped into an Error before capture.
-        expect(captureExceptionMock).toHaveBeenCalledWith(
-            expect.any(Error),
-            expect.objectContaining({ context: 'player-queue-error' }),
-        )
-    })
-
-    it('wraps a non-Error top-level player payload before capturing it', () => {
-        const { playerHandlers } = createPlayerWithHandlers()
-
-        ;(playerHandlers.error as TopLevelErrorHandler)('raw failure' as any)
-
-        expect(captureExceptionMock).toHaveBeenCalledWith(
-            expect.any(Error),
-            expect.objectContaining({ context: 'player-unhandled-error' }),
-        )
-    })
-
-    it('skips when stream recovery has no requester, no tracks, no alternative, or no current track', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        const streamError = new Error('Could not extract stream')
-
-        const queueNoRequester = {
-            guild: { id: 'guild-a', name: 'Guild A' },
-            metadata: {},
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Song A',
-            },
-            player: { search: jest.fn() },
-            node: {
-                isPlaying: jest.fn(() => false),
-                play: jest.fn(),
-                skip: jest.fn(),
-            },
-            removeTrack: jest.fn(),
-            addTrack: jest.fn(),
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queueNoRequester as any,
-            streamError,
-        )
-        await flushPromises()
-        expect(queueNoRequester.node.skip).toHaveBeenCalled()
-        expect(queueNoRequester.player.search).not.toHaveBeenCalled()
-
-        const queueNoTracks = {
-            ...queueNoRequester,
-            metadata: { requestedBy: { id: 'user-a' } },
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Song B',
-                requestedBy: { id: 'user-a' },
-            },
-            player: { search: jest.fn().mockResolvedValue({ tracks: [] }) },
-            node: {
-                isPlaying: jest.fn(() => false),
-                play: jest.fn(),
-                skip: jest.fn(),
-            },
-            removeTrack: jest.fn(),
-            addTrack: jest.fn(),
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queueNoTracks as any,
-            streamError,
-        )
-        await flushPromises()
-        expect(queueNoTracks.node.skip).toHaveBeenCalled()
-
-        const queueNoAlternative = {
-            ...queueNoTracks,
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Song C',
-                requestedBy: { id: 'user-a' },
-            },
-            player: {
-                search: jest.fn().mockResolvedValue({
-                    tracks: [{ url: 'https://example.com/current' }],
-                }),
-            },
-            node: {
-                isPlaying: jest.fn(() => true),
-                play: jest.fn(),
-                skip: jest.fn(),
-            },
-            removeTrack: jest.fn(),
-            addTrack: jest.fn(),
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queueNoAlternative as any,
-            streamError,
-        )
-        await flushPromises()
-        expect(queueNoAlternative.node.skip).toHaveBeenCalled()
-
-        const queueNoCurrentTrack = {
-            ...queueNoTracks,
-            currentTrack: null,
-            node: {
-                isPlaying: jest.fn(() => false),
-                play: jest.fn(),
-                skip: jest.fn(),
-            },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queueNoCurrentTrack as any,
-            streamError,
-        )
-        await flushPromises()
-        expect(queueNoCurrentTrack.node.skip).toHaveBeenCalled()
-    })
-
-    it('logs and swallows unexpected failures inside player error handling', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        providerFromTrackMock.mockImplementationOnce(() => {
-            throw new Error('provider lookup failed')
-        })
-        const queue = {
-            guild: { id: 'guild-x', name: 'Guild X' },
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Song X',
-            },
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(errorLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: 'Error in player error handler:',
-                data: expect.objectContaining({
-                    errorMessage: 'provider lookup failed',
-                }),
-            }),
-        )
-    })
-
-    it('sends Discord error embed when YouTube recovery finds no tracks', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        providerFromTrackMock.mockReturnValue('youtube')
-        analyzeYouTubeErrorMock.mockReturnValue({ isParserError: false })
-
-        const channelSendMock = jest.fn().mockResolvedValue({})
-        const queue = {
-            guild: { id: 'guild-notify', name: 'Guild Notify' },
-            metadata: {
-                requestedBy: { id: 'user-1' },
-                channel: { id: 'ch-1', send: channelSendMock },
-            },
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Unavailable Song',
-                requestedBy: { id: 'user-1' },
-            },
-            player: { search: jest.fn().mockResolvedValue({ tracks: [] }) },
-            insertTrack: jest.fn(),
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(channelSendMock).toHaveBeenCalledWith(
-            expect.objectContaining({ embeds: expect.any(Array) }),
-        )
-        expect(queue.node.skip).toHaveBeenCalled()
-        expect(warnLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: expect.stringContaining(
-                    'YouTube recovery found nothing',
-                ),
-            }),
-        )
-    })
-
-    it('logs Discord notification failures during stream recovery', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        providerFromTrackMock.mockReturnValue('youtube')
-        analyzeYouTubeErrorMock.mockReturnValue({ isParserError: false })
-
-        const sendError = new Error('missing access')
-        const queue = {
-            guild: { id: 'guild-notify-fail', name: 'Guild Notify Fail' },
-            metadata: {
-                requestedBy: { id: 'user-1' },
-                channel: {
-                    id: 'ch-1',
-                    send: jest.fn().mockRejectedValue(sendError),
-                },
-            },
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Unavailable Song',
-                requestedBy: { id: 'user-1' },
-            },
-            player: { search: jest.fn().mockResolvedValue({ tracks: [] }) },
-            insertTrack: jest.fn(),
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(debugLogMock).toHaveBeenCalledWith({
-            message: 'Failed to notify channel about stream failure',
-            error: sendError,
-            data: {
-                guildId: 'guild-notify-fail',
-                trackTitle: 'Unavailable Song',
-            },
-        })
-        expect(queue.node.skip).toHaveBeenCalled()
-    })
-
-    it('sends Discord error embed when no requestedBy user', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        providerFromTrackMock.mockReturnValue('youtube')
-        analyzeYouTubeErrorMock.mockReturnValue({ isParserError: false })
-
-        const channelSendMock = jest.fn().mockResolvedValue({})
-        const queue = {
-            guild: { id: 'guild-no-user', name: 'Guild No User' },
-            metadata: {
-                requestedBy: null,
-                channel: { id: 'ch-1', send: channelSendMock },
-            },
-            currentTrack: {
-                url: 'https://example.com/current',
-                title: 'Song No User',
-                requestedBy: null,
-            },
-            player: { search: jest.fn() },
-            insertTrack: jest.fn(),
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(channelSendMock).toHaveBeenCalled()
-        expect(queue.node.skip).toHaveBeenCalled()
-    })
-
-    it('skips and notifies when YouTube search times out (Promise.race null)', async () => {
-        // Search returns null (simulates timeout resolution) — recovery must
-        // treat it the same as no results.
-        const { queueHandlers } = createPlayerWithHandlers()
-        providerFromTrackMock.mockReturnValue('youtube')
-        analyzeYouTubeErrorMock.mockReturnValue({ isParserError: false })
-
-        const channelSendMock = jest.fn().mockResolvedValue({})
-        const queue = {
-            guild: { id: 'guild-timeout', name: 'Guild Timeout' },
-            metadata: {
-                requestedBy: { id: 'user-1' },
-                channel: { id: 'ch-1', send: channelSendMock },
-            },
-            currentTrack: {
-                url: 'https://example.com/track',
-                title: 'Slow Search Song',
-                requestedBy: { id: 'user-1' },
-            },
-            // search returns null — simulates the timeout branch resolving null
-            player: { search: jest.fn().mockResolvedValue(null) },
-            insertTrack: jest.fn(),
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Could not extract stream'),
-        )
-        await flushPromises()
-
-        expect(queue.node.skip).toHaveBeenCalled()
-        expect(channelSendMock).toHaveBeenCalled()
-    })
-
-    it('skips immediately without searching when currentTrack.title is empty', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        analyzeYouTubeErrorMock.mockReturnValue({ isParserError: false })
-
-        const channelSendMock = jest.fn().mockResolvedValue({})
-        const searchMock = jest.fn()
-        const queue = {
-            guild: { id: 'guild-1', name: 'Guild 1' },
-            metadata: {
-                requestedBy: { id: 'user-1' },
-                channel: { id: 'ch-1', send: channelSendMock },
-            },
-            currentTrack: {
-                url: 'https://www.youtube.com/watch?v=abc',
-                title: '',
-                requestedBy: { id: 'user-1' },
-            },
-            player: { search: searchMock },
-            insertTrack: jest.fn(),
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Bridge exhausted: no stream for empty title'),
-        )
-        await flushPromises()
-
-        expect(searchMock).not.toHaveBeenCalled()
-        expect(queue.node.skip).toHaveBeenCalled()
-        expect(warnLogMock).toHaveBeenCalledWith(
-            expect.objectContaining({
-                message: expect.stringContaining('no title'),
-            }),
-        )
-    })
-
-    it('falls back to console.error with the original message, error and data when errorLog itself throws', () => {
-        const { playerHandlers } = createPlayerWithHandlers()
-        const loggerError = new Error('errorLog transport failed')
-        errorLogMock.mockImplementationOnce(() => {
-            throw loggerError
-        })
-        const consoleErrorSpy = jest
-            .spyOn(console, 'error')
-            .mockImplementation(() => {})
-
-        const originalError = new Error('Unhandled player error')
-        ;(playerHandlers.error as TopLevelErrorHandler)(originalError)
-
-        expect(consoleErrorSpy).toHaveBeenCalledWith(
-            'Unhandled player error:',
-            originalError,
-            expect.objectContaining({
-                errorMessage: 'Unhandled player error',
-                errorName: 'Error',
-            }),
-            loggerError,
-        )
-
-        consoleErrorSpy.mockRestore()
-    })
-
-    it('uses "this track" in notification embed when trackTitle is empty', async () => {
-        const { queueHandlers } = createPlayerWithHandlers()
-        analyzeYouTubeErrorMock.mockReturnValue({ isParserError: false })
-
-        const queue = {
-            guild: { id: 'guild-1', name: 'Guild 1' },
-            metadata: {
-                requestedBy: { id: 'user-1' },
-                channel: { id: 'ch-1', send: jest.fn().mockResolvedValue({}) },
-            },
-            currentTrack: {
-                url: 'https://www.youtube.com/watch?v=abc',
-                title: '',
-                requestedBy: { id: 'user-1' },
-            },
-            player: { search: jest.fn() },
-            insertTrack: jest.fn(),
-            node: { skip: jest.fn() },
-        }
-
-        ;(queueHandlers.playerError as PlayerErrorHandler)(
-            queue as any,
-            new Error('Bridge exhausted: no stream for empty title'),
-        )
-        await flushPromises()
-        await flushPromises()
-
-        expect(createErrorEmbedMock).toHaveBeenCalledWith(
-            expect.stringContaining('Could not play'),
-            expect.stringContaining('this track'),
-        )
-    })
-})
